Change text rendering to use floating point positions

+++ Renderer/src/main/kotlin/app/ultradev/hytaleuiparser/renderer/element/impl/UILabelElement.kt
@@ -22,10 +22,10 @@ class UILabelElement(
         val wrapped = textRenderInfo.wrap(text, available.width)
 
         val height = textRenderInfo.getHeight() * wrapped.size
-        val width = wrapped.maxOfOrZero { textRenderInfo.getWidth(it) }
+        val width = wrapped.maxOfOrZero { textRenderInfo.getWidth(it).toInt() }
         return BoxSize(
             width,
-            height
+            height.toInt()
         )
     }
