+++ Renderer/src/main/kotlin/app/ultradev/hytaleuiparser/renderer/element/impl/UITextButtonElement.kt
@@ -23,8 +23,8 @@ class UITextButtonElement(
 
     override fun contentDesiredSize(available: BoxSize): BoxSize {
         return BoxSize(
-            textRenderInfo.getHeight(),
-            textRenderInfo.getWidth(text)
+            textRenderInfo.getHeight().toInt(),
+            textRenderInfo.getWidth(text).toInt()
         )
     }
